typo: fix misspelling of TARGETING
While the spelling "targetting" is a valid, yet obsolete, British one, this
code base was likely documented, at the time, for a predominantly American
audience. Thus the spelling might have been intended to be TARGETING instead
of TARGETTING. In all fairness, this is perhaps, a rather recentist reading
of the comments.

Signed-off-by: Eden G. Adogla <eden@edengrail.io>