Docker build of Apollo-11 sources.
This is a first attempt at a Docker configuration for a build
of the Apollo-11 sources using the VirtualAgc tools.

To use it, you'll need Docker (on Linux, Mac or Windows). Then do:
  $ docker build -t apollo-11 .

It first clones and builds the VirtualAgc tools and mission code.
Then it compiles the Apollo-11 sources (from the current directory)
and compares the generated rope binaries against those in the
VirtualAgc build.

This is based on an Ubuntu 16.04 base image.

Note that the VirtualAgc tools are sensitive to spaces/tabs, so
assembly errors may result from using a different formatting
standard.